http://localhost:8888/notebooks/Mercedes%20-%20Benz%20project.ipynb#
You are required to reduce the time that cars spend on the test bench. You will work with a dataset representing different permutations of features in a Mercedes-Benz car to predict the time it takes to pass testing. Optimal algorithms will contribute to faster testing, resulting in lower carbon dioxide emissions without reducing Daimler’s standards.